requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.1
numpy==1.24.3
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
aiohttp==3.9.1
openai==1.3.0 